+++ effectful/tests/EnvTests.hs
@@ -23,7 +23,9 @@ envTests = testGroup "Env"
   , testCase "unsafeCoerce doesn't work" test_noUnsafeCoerce
   , testCase "interpose works" test_interpose
   , testCase "interpose/provider works" test_interposeProvider
-  , testCase "borrow/lend works" test_borrowLend
+  , testCase "localSeqLend/localSeqBorrow works" test_lendBorrowSeparate
+  , testCase "localLendBorrow works" test_lendBorrowCombined
+  , testCase "lend/borrow combined shares storage" test_lendBorrowForkShares
   ]
 
 test_tailEnv :: Assertion
@@ -200,32 +202,80 @@ doubleB = interpose_ $ \case
 
 ----------------------------------------
 
-test_borrowLend :: Assertion
-test_borrowLend = runEff $ do
-  runX 1 2 . evalState @[Int] [3] . runReader () . runReader @[Int] [4] $ do
+test_lendBorrowSeparate :: Assertion
+test_lendBorrowSeparate = runEff $ do
+  runX1 1 2 . evalState @[Int] [3] . runReader () . runReader @[Int] [4] $ do
+    U.assertEqual "expected result" [1,2,3,4,1,2,3,4] =<< send X
+
+test_lendBorrowCombined :: Assertion
+test_lendBorrowCombined = runEff $ do
+  runX2 1 2 . evalState @[Int] [3] . runReader () . runReader @[Int] [4] $ do
     U.assertEqual "expected result" [1,2,3,4,1,2,3,4] =<< send X
 
 data X :: Effect where
   X :: (State [Int] :> es, Reader [Int] :> es, Reader () :> es) => X (Eff es) [Int]
 type instance DispatchOf X = Dynamic
 
-runX :: Int -> Int -> Eff (X : es) a -> Eff es a
-runX s0 r0 = reinterpret (evalState s0 . evalState () . runReader r0) $ \env -> \case
+runX1 :: Int -> Int -> Eff (X : es) a -> Eff es a
+runX1 s0 r0 = reinterpret (evalState s0 . evalState () . runReader r0) $ \env -> \case
   X -> localSeqUnlift env $ \unlift -> do
-    as <- localSeqLend @[State Int, Reader Int] env $ \withHandlerEffs -> do
-      unlift . withHandlerEffs $ do
-        () <- ask
-        s <- get @Int
-        r <- ask @Int
-        ss <- get @[Int]
-        rs <- ask @[Int]
-        pure $ [s, r] ++ ss ++ rs
-    bs <- localSeqBorrow @[Reader [Int], State [Int], Reader ()] env $ \withEffs -> do
-      withEffs $ do
-        () <- ask
-        s <- get @Int
-        r <- ask @Int
-        ss <- get @[Int]
-        rs <- ask @[Int]
-        pure $ [s, r] ++ ss ++ rs
-    pure $ as ++ bs
+    localSeqLend @[State Int, Reader Int] env $ \lend -> do
+      localSeqBorrow @[Reader [Int], State [Int], Reader ()] env $ \borrow -> do
+        as <- unlift . lend $ do
+          () <- ask
+          s <- get @Int
+          r <- ask @Int
+          ss <- get @[Int]
+          rs <- ask @[Int]
+          pure $ [s, r] ++ ss ++ rs
+        bs <- borrow $ do
+          () <- ask
+          s <- get @Int
+          r <- ask @Int
+          ss <- get @[Int]
+          rs <- ask @[Int]
+          pure $ [s, r] ++ ss ++ rs
+        pure $ as ++ bs
+
+runX2 :: Int -> Int -> Eff (X : es) a -> Eff es a
+runX2 s0 r0 = reinterpret (evalState s0 . evalState () . runReader r0) $ \env -> \case
+  X -> localSeqUnlift env $ \unlift -> do
+    localLendBorrow
+      @[State Int, Reader Int]
+      @[Reader [Int], State [Int], Reader ()]
+      env SeqUnlift $ \lend borrow -> do
+        as <- unlift . lend $ do
+          () <- ask
+          s <- get @Int
+          r <- ask @Int
+          ss <- get @[Int]
+          rs <- ask @[Int]
+          pure $ [s, r] ++ ss ++ rs
+        bs <- borrow $ do
+          () <- ask
+          s <- get @Int
+          r <- ask @Int
+          ss <- get @[Int]
+          rs <- ask @[Int]
+          pure $ [s, r] ++ ss ++ rs
+        pure $ as ++ bs
+
+-- | Under 'SeqForkUnlift' state modification done via the borrowing function
+-- must also be visible via the lending function (they share the forked
+-- storage), but invisible to the handler.
+test_lendBorrowForkShares :: Assertion
+test_lendBorrowForkShares = runEff . evalState @Int 0 . runY $ do
+  U.assertEqual "lend sees borrow's write" 1 =<< send Y
+  U.assertEqual "handler is isolated" 0 =<< get @Int
+
+data Y :: Effect where
+  Y :: State Int :> es => Y (Eff es) Int
+type instance DispatchOf Y = Dynamic
+
+runY :: State Int :> es => Eff (Y : es) a -> Eff es a
+runY = interpret $ \env Y ->
+  localLendBorrow @'[State Int] @'[State Int] env SeqForkUnlift $ \lend borrow -> do
+    -- Write via borrow (runs against the forked clone)...
+    borrow $ put @Int 1
+    -- ...and read it back via lend (against the same forked clone).
+    localSeqUnlift env $ \unlift -> unlift . lend $ get @Int